De-<script>-ified tst.
<script> is an optional tentative controversial zpt feature that
sessions should not depend on.

In addition, exceptions are something not well handled in untrusted
python code. We will soon disable try/except in untrusted python code
until we have figured out how to do it.

Finally, the zpt session support should should move into a separate
package, which could be a sub-package of session.  The reason is that
zpt is, theoretically, an optional feature. Someone might want to have
sessions without having zpt.